Refactor build configuration and update dependencies
This commit updates the Android build configuration by replacing the direct inclusion of the Google Services plugin with an alias in the `build.gradle.kts` files. It also modifies the application name in the output file naming convention, changes the annotation processor from `kapt` to `ksp` for Room dependencies, and improves the formatting of the AndroidManifest.xml. Additionally, the version of the Google Services plugin is now referenced in the `libs.versions.toml` file.