MA-311: Refactor AppStyledCard component and update PagerCard to use AppStyledCard
The AppStyledCard component was refactored to use a Box with a background modifier instead of a Surface for more direct control over styling. The PagerCard component was updated to use the new AppStyledCard component.

The AppText component was also updated to remove the modifier parameter.

The MeAppTheme was updated to include AppTypography and LocalTypography.