Refactor: Remove unused functions from DateTimeUtil
This commit streamlines the `DateTimeUtil` object by removing a large number of unused date and time formatting and conversion functions.

The following functions have been removed:
- `getTodayFormattedDate`
- `getFormattedDate`
- `getFormattedDayWithTime`
- `getFormattedMonthDay`
- `getFormattedTime`
- `getMonth`
- `getMonthYear`
- `getMonthDayYear`
- `getYear`
- `getDay`
- `getCurrentDatetimeIsoString`
- `getDatetimeIsoString`
- `getCurrentTimeWithTimeZone`
- `getIntervalDatetimeIsoString`
- `getDateStringFormattedForDatePicker`
- `getMaxBirthdayOffsetForDatePicker`
- `getBirthdayFormattedString`
- `getUserTimezone`
- `getUserTimezoneOffset`
- `getTimeZoneInMinutes`
- `getDateFormatFromMilliseconds`
- `getEpochMillisFromDateString`

The only remaining functions are `getCurrentFormattedTimestamp` and `getMinBirthdayOffsetForDatePicker`. Additionally, the logic in `getMinBirthdayOffsetForDatePicker` was slightly adjusted to set the time to the end of the day.