MA-1692 - Save scale immediately upon discovery
This commit modifies the `LcbtBLESetupViewModel` to ensure that a discovered scale is saved to the `deviceService` immediately after being discovered, regardless of its connection status.

Previously, the scale was only saved if its `connectionStatus` was `CONNECTED`. This condition has been removed to allow saving the scale information as soon as it is available.